Create the HSV to RGB conversion

+++ code/headers/hsv_s.hpp
@@ -8,7 +8,7 @@ namespace r2d2::led_strip {
      *  Struct which contains the hue, saturation and value
      */
     struct hsv_s {
-        uint8_t hue;
+        uint16_t hue;
 	    uint8_t saturation;
         uint8_t value;
     };
